requirements.txt
flask
gunicorn
requests
flask_cors
openai
python-dotenv
groq
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.0.0
fastapi
uvicorn
python-multipart
httpx